backend/requirements.txt

werkzeug==2.2.3
flask==2.1.1
flask-cors==3.0.10

pandas==2.3.2


openai==0.27.9
python-dotenv==0.21.1
pandas==2.3.2
